Un-inlined MockApnsServerHandler, too.
